refactor(emitter): split DTS type inference helpers
