+++ crates/tsz-emitter/src/declaration_emitter/helpers/type_inference_parts/part4.rs
@@ -1,0 +1,151 @@
+use super::*;
+
+impl<'a> DeclarationEmitter<'a> {
+    pub(in crate::declaration_emitter) fn format_reused_call_structural_return_type_text(
+        &self,
+        type_text: &str,
+    ) -> String {
+        if !type_text.contains(" & ") || !type_text.contains("=> {") {
+            return type_text.to_string();
+        }
+
+        let mut out = String::with_capacity(type_text.len() + 16);
+        let mut rest = type_text;
+        let member_indent = "    ".repeat((self.indent_level + 1) as usize);
+        let closing_indent = "    ".repeat(self.indent_level as usize);
+
+        while let Some(start) = rest.find("=> {") {
+            let (before, after_marker) = rest.split_at(start + 4);
+            out.push_str(before);
+            let Some(end) = after_marker.find('}') else {
+                out.push_str(after_marker);
+                return out;
+            };
+            let body = after_marker[..end].trim();
+            if body.is_empty()
+                || body.contains('\n')
+                || body.contains(';')
+                || body.contains(',')
+                || !body.contains(':')
+            {
+                out.push_str(&after_marker[..=end]);
+                rest = &after_marker[end + 1..];
+                continue;
+            }
+
+            let member = body.trim_end_matches(';').trim();
+            out.push('\n');
+            out.push_str(&member_indent);
+            out.push_str(member);
+            out.push(';');
+            out.push('\n');
+            out.push_str(&closing_indent);
+            out.push('}');
+            rest = &after_marker[end + 1..];
+        }
+
+        out.push_str(rest);
+        out
+    }
+
+    pub(in crate::declaration_emitter) fn preserve_literal_mapped_return_type_substitutions(
+        &self,
+        source_arena: &NodeArena,
+        parameters: &NodeList,
+        call: &tsz_parser::parser::node::CallExprData,
+        type_param_names: &[String],
+        substitutions: &mut Vec<(String, String)>,
+    ) {
+        let Some(args) = call.arguments.as_ref() else {
+            return;
+        };
+
+        for (&param_idx, &arg_idx) in parameters.nodes.iter().zip(args.nodes.iter()) {
+            let Some(param_node) = source_arena.get(param_idx) else {
+                continue;
+            };
+            let Some(param) = source_arena.get_parameter(param_node) else {
+                continue;
+            };
+            let Some(param_type_text) = self
+                .emit_type_node_text_from_arena(source_arena, param.type_annotation)
+                .or_else(|| self.source_slice_from_arena(source_arena, param.type_annotation))
+            else {
+                continue;
+            };
+            let param_type_text = param_type_text.trim();
+            if !type_param_names
+                .iter()
+                .any(|name| name.as_str() == param_type_text)
+            {
+                continue;
+            }
+            let Some(substitution_text) = self
+                .enclosing_parameter_type_annotation_text_for_identifier(arg_idx)
+                .or_else(|| self.reference_declared_type_annotation_text(arg_idx))
+                .filter(|text| Self::simple_type_reference_name(text).is_some())
+                .or_else(|| self.const_literal_initializer_text(arg_idx))
+            else {
+                continue;
+            };
+            if let Some((_, existing)) = substitutions
+                .iter_mut()
+                .find(|(name, _)| name.as_str() == param_type_text)
+            {
+                *existing = substitution_text;
+            } else {
+                substitutions.push((param_type_text.to_string(), substitution_text));
+            }
+        }
+    }
+
+    pub(in crate::declaration_emitter) fn enclosing_parameter_type_annotation_text_for_identifier(
+        &self,
+        arg_idx: NodeIndex,
+    ) -> Option<String> {
+        let arg_name = self.get_identifier_text(arg_idx)?;
+        let mut current = arg_idx;
+        for _ in 0..32 {
+            let parent_idx = self.arena.parent_of(current)?;
+            let parent_node = self.arena.get(parent_idx)?;
+            if let Some(func) = self.arena.get_function(parent_node) {
+                for &param_idx in &func.parameters.nodes {
+                    let param_node = self.arena.get(param_idx)?;
+                    let param = self.arena.get_parameter(param_node)?;
+                    if self.get_identifier_text(param.name).as_deref() == Some(arg_name.as_str()) {
+                        return self
+                            .type_annotation_text_from_arena_node(self.arena, param.type_annotation)
+                            .or_else(|| {
+                                self.source_slice_from_arena(self.arena, param.type_annotation)
+                            })
+                            .map(|text| text.trim().to_string());
+                    }
+                }
+                return None;
+            }
+            current = parent_idx;
+        }
+        None
+    }
+
+    pub(in crate::declaration_emitter) fn ensure_single_line_type_literal_member_semicolon(
+        type_text: &str,
+    ) -> String {
+        let trimmed = type_text.trim();
+        if trimmed.contains('\n') {
+            return type_text.to_string();
+        }
+        let Some(inner) = trimmed
+            .strip_prefix('{')
+            .and_then(|text| text.strip_suffix('}'))
+            .map(str::trim)
+        else {
+            return type_text.to_string();
+        };
+        if inner.is_empty() || inner.ends_with(';') || inner.contains(';') || !inner.contains(':') {
+            type_text.to_string()
+        } else {
+            format!("{{ {inner}; }}")
+        }
+    }
+}
